Moved query post request into separate component TextService and moved replaceAll regular expression into separate component TextService
