+++ ui/src/elle/tools/wordanalyser/WordAnalyser.js
@@ -1,10 +1,10 @@
-import {memo, useContext, useEffect, useRef, useState} from 'react';
-import {Input} from './textinput/Input';
-import {WordInfo} from './WordInfo';
+import { memo, useContext, useEffect, useRef, useState } from 'react';
+import { Input } from './textinput/Input';
+import { WordInfo } from './WordInfo';
 import './styles/WordAnalyser.css';
 import TextUpload from '../../components/TextUpload';
-import {Alert, Box, Fade, Grid, IconButton, Typography} from '@mui/material';
-import {useTranslation} from 'react-i18next';
+import { Alert, Box, Fade, Grid, IconButton, Typography } from '@mui/material';
+import { useTranslation } from 'react-i18next';
 import '../../translations/i18n';
 import i18n from 'i18next';
 import {
@@ -18,8 +18,8 @@ import {
   WordContext
 } from './Contexts';
 import CloseIcon from '@mui/icons-material/Close';
-import {queryStore} from "../../store/QueryStore";
-import {loadFetch} from "../../service/LoadFetch";
+import { queryStore } from '../../store/QueryStore';
+import { toolsPopulatePostQuery } from '../../service/TextService';
 
 function WordAnalyser() {
   const [analysedInput, setAnalysedInput] = useContext(AnalyseContext);
@@ -37,41 +37,20 @@ function WordAnalyser() {
   const {t} = useTranslation();
   const [open, setOpen] = useState(false);
   const [border, setBorder] = useState(0);
-  const [storeData, setStoreData] = useState()
-  const inputRef = useRef()
+  const [storeData, setStoreData] = useState();
+  const inputRef = useRef();
 
   useEffect(() => {
-    postRequest();
+    toolsPopulatePostQuery(setStoreData);
   }, []);
 
   useEffect(() => {
-    setTextFromFile(storeData)
+    setTextFromFile(storeData);
   }, [storeData]);
 
   queryStore.subscribe(() => {
-    postRequest();
+    toolsPopulatePostQuery(setStoreData);
   });
-
-  const postRequest = () => {
-    const queryStoreState = queryStore.getState();
-    if (queryStoreState.corpusTextIds) {
-      loadFetch('/api/texts/kysitekstid', {
-        method: 'POST',
-        body: JSON.stringify({ids: queryStoreState.corpusTextIds}),
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      }).then(res => res.text())
-        .then(result => {
-          if (queryStoreState.ownTexts) {
-            result = result.concat(" ", queryStoreState.ownTexts)
-          }
-          setStoreData(result.replaceAll('\\n\\n', ' ').replaceAll('\\n', ' ').replaceAll('&quot;', '"'))
-        });
-    } else if (queryStoreState.ownTexts) {
-      setStoreData(queryStoreState.ownTexts);
-    }
-  };
 
   // get words
   const getWords = async (input) => {
