+++ ui/src/elle/util/TextUtils.js
@@ -1,0 +1,3 @@
+export const regExpReplaceAll = (text) => {
+  return text.replaceAll('\\n\\n', ' ').replaceAll('\\n', ' ').replaceAll('&quot;', '"');
+};
